Add typeid to Shape struct
This commit adds a `typeid` field to the `Shape` struct to store the `TypeId` of the represented type. It updates all implementations of `Shapely` to populate this new field using the `mini_typeid` module. The `Shape` struct is also updated to derive `PartialEq`, `Eq`, and `Hash` based on the `typeid` field.